+++ packages/components/src/components/multi-select/multi-select/multi-select-styles.ts
@@ -35,7 +35,6 @@ import {
   formElementPaddingVertical,
   getCalculatedFormElementPaddingHorizontal,
 } from '../../../styles/form-styles';
-import { INTERNAL_MULTI_SELECT_SLOT } from './multi-select-utils';
 
 export const getComponentCss = (
   direction: SelectDropdownDirectionInternal,
@@ -43,7 +42,6 @@ export const getComponentCss = (
   isDisabled: boolean,
   hideLabel: BreakpointCustomizable<boolean>,
   state: FormState,
-  isWithinForm: boolean,
   isNativePopoverCase: boolean,
   theme: Theme
 ): string => {
@@ -60,15 +58,6 @@ export const getComponentCss = (
         }),
       },
       ...preventFoucOfNestedElementsStyles,
-      ...(isWithinForm &&
-        addImportantToEachRule({
-          [`::slotted([slot=${INTERNAL_MULTI_SELECT_SLOT}])`]: {
-            position: 'absolute',
-            opacity: 0,
-            height: '0px',
-            bottom: 0,
-          },
-        })),
       // TODO: re-use select-wrapper-style
       input: getInputStyles(isDisabled, direction, isOpen, state, theme),
     },
@@ -77,9 +66,6 @@ export const getComponentCss = (
       gap: spacingStaticXSmall,
       // min width is needed for showing at least 1 character in very narrow containers. The "1rem" value is the minimum safe zone to show at least 1 character plus the ellipsis dots.
       minWidth: `calc(1rem + ${formElementPaddingHorizontal} + ${borderWidthBase} * 2 + ${getCalculatedFormElementPaddingHorizontal(2)})`,
-      ...(isWithinForm && {
-        position: 'relative', // Necessary for native HTML validation box positioning (internal-select)
-      }),
     },
     wrapper: {
       position: 'relative',